fix: multiarch builds and multiple repos using buildx
Refactors `Makefile.docker` to use buildx and to support multiple repos.

`docker buildx build` removes the need to push multiple tags, use
`docker manifest create` and delete tags so that all arches appear on a
single tag. This was not feasible to complete when adding ghcr.io hosted
images.

Updates to Dockerfile to use TARGETPLATFORM to pickup the correct
binary.

Updates to E2E tests for new target and test mode.

ref: #6661

fstuff